[VVC][Coverity] Fix INTEGER_OVERFLOW CID 3583108 in VVCScanOrderTables::InitScanOrderTables
Change loop coordinate variables x and y from uint32_t to int32_t in the
SCAN_DIAG initialization loop to eliminate false-positive INTEGER_OVERFLOW
Coverity warnings (CID 3583108).

Coverity flagged two occurrences:
- Line 111: y-- underflow when y == 0 (uint32_t)
- Line 100: y += x + 1 overflow

These are false positives: the else branch (y--) requires y != 0 as guaranteed
by the outer condition, and block dimensions (max 128x128) keep y and x far
below int32_t/uint32_t overflow thresholds. Using signed int32_t for the
coordinate variables makes the intent clear to static analysis tools.

No behavioral change; x and y are always non-negative and bounded by blkWidth
and blkHeight respectively during correct execution.

Coverity: CID 3583108

+++ _studio/shared/umc/codec/vvc_dec/src/umc_vvc_bitstream_headers.cpp
@@ -87,7 +87,7 @@ namespace UMC_VVC_DECODER
                 m_scanOrders[SCAN_UNGROUPED][SCAN_DIAG][blkWidthIdx][blkHeightIdx] = scan;
                 ScanElement* scanEnd = scan + totalValues;
 
-                uint32_t x = 0, y = 0;
+                int32_t x = 0, y = 0;
                 scan->idx = 0;
                 scan->x = 0;
                 scan->y = 0;
@@ -95,15 +95,15 @@ namespace UMC_VVC_DECODER
 
                 while (scan < scanEnd)
                 {
-                    if (y == 0 || x == blkWidth - 1)
+                    if (y == 0 || x == (int32_t)blkWidth - 1)
                     {
                         y += x + 1;
                         x = 0;
 
-                        if (y >= blkHeight)
+                        if (y >= (int32_t)blkHeight)
                         {
-                            x += y - (blkHeight - 1);
-                            y = blkHeight - 1;
+                            x += y - ((int32_t)blkHeight - 1);
+                            y = (int32_t)blkHeight - 1;
                         }
                     }
                     else
@@ -112,9 +112,9 @@ namespace UMC_VVC_DECODER
                         x++;
                     }
 
-                    scan->idx = y * blkWidth + x;
-                    scan->x = x;
-                    scan->y = y;
+                    scan->idx = (uint32_t)(y * (int32_t)blkWidth + x);
+                    scan->x = (uint32_t)x;
+                    scan->y = (uint32_t)y;
                     scan++;
                 }
             }
